Adds Support for Breakfast, Lunch, Dinner, Notes, and Slow Cookers
Instead of specifying a meal, you specify a breakfast, lunch, and dinner
on the day. If your day has more than one meal, it outputs the type
of the meal.

Notes can be displayed on a day instead of a meal and shows in italics.

A slow cooker alert can be added so you know you don't cook at your
normal time.

Adds meal plan for 06282020.